Refactor, remove hardcoded links
